+++ src/main/java/com/historycode/ui/component/BaseComponent.java
@@ -1,28 +1,20 @@
 package com.historycode.ui.component;
 
 import com.historycode.ui.Base;
+import lombok.Getter;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.pagefactory.DefaultElementLocatorFactory;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 public abstract class BaseComponent extends Base {
 
+    @Getter
     protected WebElement rootElement;
 
     public BaseComponent(WebDriver driver, WebElement rootElement) {
         super(driver);
         PageFactory.initElements(new DefaultElementLocatorFactory(rootElement), this);
         this.rootElement = rootElement;
-    }
-
-    public BaseComponent(WebDriver driver) {
-        super(driver);
-    }
-
-    public void waitUntilElementVisible(WebElement element) {
-        sleep(2000);
-        wait.until(ExpectedConditions.visibilityOf(element));
     }
 }
