+++ src/main/java/com/historycode/ui/component/header/HeaderComponent.java
@@ -1,6 +1,7 @@
 package com.historycode.ui.component.header;
 
 import com.historycode.ui.component.BaseComponent;
+import com.historycode.ui.component.BurgerMenu.BurgerMenuComponent;
 import lombok.Getter;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -23,8 +24,24 @@ public class HeaderComponent extends BaseComponent {
     @Getter
     SearchElement searchElement;
 
+
+    @FindBy(xpath = "//div[contains(@class, 'rightPartContainer')]//div[contains(@class, 'drawerContainer')]//div")
+    private WebElement burgerMenuBtn;
+    @FindBy(xpath = "//div[@class='ant-drawer-body']")
+    private WebElement burgerMenuBody;
+
     public HeaderComponent(WebDriver driver, WebElement rootElement) {
         super(driver, rootElement);
         searchElement = new SearchElement(driver, searchNode);
+    }
+
+    public boolean isBurgerMenuVisible() {
+        return burgerMenuBody.isDisplayed();
+    }
+
+
+    public BurgerMenuComponent clickBurgerMenuBtn(){
+        burgerMenuBtn.click();
+        return new BurgerMenuComponent(driver, burgerMenuBody);
     }
 }
