+++ src/main/java/com/historycode/ui/page/streetCodePage/components/ChronologyFilmCardComponent.java
@@ -35,12 +35,6 @@ public class ChronologyFilmCardComponent extends BaseComponent {
 
     public ChronologyFilmCardComponent(WebDriver driver, WebElement rootElement) {
         super(driver, rootElement);
-        PageFactory.initElements(driver, this);
-    }
-
-    public ChronologyFilmCardComponent(WebDriver driver) {
-        super(driver);
-        PageFactory.initElements(driver, this);
     }
 
     public WebElement getFilmCardByIndex(int index) {
